feat(nvim): add git/local folder fuzzy find files
**Summary:**
Similar to the behavior we had with fzf, we want to search in all git
files when in a path "managed" by git or we want to look in the current
folder and subfolders.

As telescope does not provide an option for this ootb, we add the
solution from https://github.com/nvim-telescope/telescope.nvim/issues/2183.